+++ src/main/java/com/study/board/domain/post/service/PostService.java
@@ -4,9 +4,9 @@ import com.study.board.domain.exception.PostEditAccessDeniedException;
 import com.study.board.domain.exception.PostNotFoundException;
 import com.study.board.domain.exception.UserNotFoundException;
 import com.study.board.domain.post.domain.Post;
-import com.study.board.domain.post.domain.PostRepository;
+import com.study.board.domain.post.repository.PostRepository;
 import com.study.board.domain.user.domain.User;
-import com.study.board.domain.user.domain.UserRepository;
+import com.study.board.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
