+++ src/main/java/com/study/board/domain/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.study.board.domain.user.domain;
+package com.study.board.domain.user.repository;
 
+import com.study.board.domain.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
